--- a/profiler/internal/fastdelta/location_index.go
+++ b/profiler/internal/fastdelta/location_index.go
@@ -1,3 +1,8 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2022 Datadog, Inc.
+
 package fastdelta
 
 // locationIndex links location IDs to the addresses, mappings, and function