--- a/profiler/internal/fastdelta/pprof_records.go
+++ b/profiler/internal/fastdelta/pprof_records.go
@@ -1,3 +1,8 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2022 Datadog, Inc.
+
 package fastdelta
 
 // enum types for pprof protobuf records