--- a/profiler/internal/fastdelta/testdata/fuzz/FuzzDelta/6281ce98ad6bb944a23bb21f7597f91ae767be28cf9ebfaaa40e3d1454c12be3
+++ b/profiler/internal/fastdelta/testdata/fuzz/FuzzDelta/6281ce98ad6bb944a23bb21f7597f91ae767be28cf9ebfaaa40e3d1454c12be3
@@ -1,0 +1,2 @@
+go test fuzz v1
+[]byte("\x12\n\xed\xef00000\b00")