--- a/profiler/internal/fastdelta/testdata/fuzz/FuzzDelta/74a0704b407ac210d9de0d409cba6bd17597ba4e67e73fcf7bdffa31438ac64f
+++ b/profiler/internal/fastdelta/testdata/fuzz/FuzzDelta/74a0704b407ac210d9de0d409cba6bd17597ba4e67e73fcf7bdffa31438ac64f
@@ -1,0 +1,2 @@
+go test fuzz v1
+[]byte("j\x1a00000000\xc9\xc9\xc9\xc9\xc9\xc9\xc9\xc9\xc9000000000")