## Description

The [pprof protobuf structure](https://github.com/google/pprof/blob/main/proto/profile.proto) when hydrated in-memory as golang structs has significant overhead in memory and allocations.  The structure has multiple nested self-referencing sub-structs that cause GC overhead to the point of dominating CPU resources for some services.

We need to keep this structure in memory for computing delta profiles between profile cycles, making our profiling client a cause for performance regressions in some cases.

This pull adds a new delta profile implementation - `fastdelta` - that uses a streaming pprof parser and low heap allocations so the entire hydrated pprof structure no longer needs to be kept in memory between profile cycles.

## Testing Done

- tested in reliability env with `DD_PROFILING_DELTA_METHOD=comparing`, no fidelity failures
- tested on production load with `DD_PROFILING_DELTA_METHOD=fastdelta`
- similar results to the synthetic benchmark (below) in tested production service and in reliability env

## Deploy Plan

1. one `dd-trace-go` release with both delta profile implementations, old algorithm as default, new algorithm enabled with `DD_PROFILING_DELTA_METHOD` env var. This is so we can continue testing internally.
2. release N+1, make `fastdelta` the default, enable old algorithm with an env var fallback if there are any issues.
3. at some point, we may remove the old algorithm entirely.

## Results

On synthetic benchmarks

- ~4x improvement in wall time
- ~10x improvement in memory overhead
- ~20x improvement in allocations

```
name                                  old time/op    new time/op    delta
FastDelta/testdata/heap.pprof-10        2.09ms ± 1%    0.52ms ± 0%  -75.00%  (p=0.008 n=5+5)
FastDelta/testdata/big-heap.pprof-10    31.6ms ± 1%     7.0ms ± 0%  -77.93%  (p=0.008 n=5+5)

name                                  old alloc/op   new alloc/op   delta
FastDelta/testdata/heap.pprof-10        2.98MB ± 0%    0.23MB ± 0%  -92.44%  (p=0.008 n=5+5)
FastDelta/testdata/big-heap.pprof-10    35.4MB ± 0%     1.7MB ± 0%  -95.18%  (p=0.008 n=5+5)

name                                  old allocs/op  new allocs/op  delta
FastDelta/testdata/heap.pprof-10         41.2k ± 0%      1.3k ± 0%  -96.74%  (p=0.008 n=5+5)
FastDelta/testdata/big-heap.pprof-10      524k ± 0%       14k ± 0%  -97.25%  (p=0.008 n=5+5)
```

## Change Summary

- vendors [fastdelta](https://github.com/nsrip-dd/fastdelta/)
- fix for `duration_nanos` tracking
- fixes for over-fit test fixtures
- test coverage for the 3 delta implementations (fastdelta, pprofile, comparing)
- comparing telemetry