@pmbauer 2022-10-19
@nsrip-dd I've addressed each comment with isolated commits with the message `review: ...` for easy comparison.  I have one more fix to make for https://github.com/DataDog/dd-trace-go/pull/1511/files#diff-2342fcb89653c2a2af7d9fda4d2e9aa2a96dacf2df3a557c8b7b18cb4a499b55R352