@nsrip-dd 2022-10-20
Okay, I think all the stuff from my first review has been addressed. Thanks Paul!

At this point, we should have Felix take a look. We'll mainly want to focus on the following things in my opinion:

* Are we satisfied with the correctness of the algorithm and its implementation?
* Are the tests we have now sufficient? We've tested this on real apps with no fidelity problems, and the performance looks really good. But we can't sustainably do several release cycles with intensive fidelity checking every time we want to further change this. I think we can articulate the risks of this approach and probably come up with a reasonable way to test against them.
  * Off the top of my head, I think the biggest risk is the input dramatically changing. For example, I hard-coded the array that holds the sample values to 4 entries, so we the sample map doesn't have to pointer-chase to dynamically-allocated slices of values. 4 values is the most that a Go profile currently has (the heap profile). What if Go adds more values to a profile (like allocation lifetimes)? We also want to worry about being too trusting of indices in the profile, and reading out of bounds. 

Any style stuff or aesthetic refactoring can wait until a subsequent PR IMO. We should at least be happy with the overall design :)