@nsrip-dd 2022-10-26
Nope, all good for me! IMO these changes are OK to merge as-is especially since this is off-by-default behind an environment variable. I'll check in with Felix tomorrow.