@pmbauer 2022-11-02
> 🐞 Maybe I'm missing something, but it seems like the entire stringsTable implementation assumes that string table ID references remain stable between profiles?

The stringTable is not used between profiles and is [reset](https://github.com/DataDog/dd-trace-go/blob/e8bfed74c149c818476da7d9644096e771c1d341/profiler/internal/fastdelta/fd.go#L143) at the start of each delta.  Unless I'm misunderstanding something, we don't assume the IDs are stable between profiles.