@felixge 2022-11-03
>  The strings that get written out to the delta pprof will have the same offsets as the input pprof passed to Delta, but there is no assumed carry-over from one delta pprof to the next.

💡 Ah, got it. For some reason I thought that there might be references into the string table from the previous profile. But I can see now that this is clearly not the case. All references are from the current profile, so it doesn't matter if the same string had a different index in the previous profile. ✅