@zdyj3170101136 2024-09-06
The delta profile I designed relies on the built-in ordering of go performance profiles, which does not consume cpu.

please see https://github.com/DataDog/dd-trace-go/issues/2804.